refactor: remove domain package from the core
