+++ app/src/main/java/com/monstarlab/core/network/OkHttpModule.kt
@@ -1,4 +1,4 @@
-package com.monstarlab.core.injection.modules
+package com.monstarlab.core.network
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import com.monstarlab.BuildConfig
